Feat: NewsArticle reliability 데이터 타입 변경 Int -> Double
Former-commit-id: 79231ec1203aa3fdc941fab27ebdecd1d9027373

+++ android/app/src/main/java/com/ssafy/firstproject/data/model/response/NewsArticle.kt
@@ -11,7 +11,7 @@ data class NewsArticle(
     val paragraphs: List<String>,
     val pubDate: String,
     val pubDateTimestamp: Long,
-    val reliability: Int,
+    val reliability: Double,
     val title: String,
     val topImage: String
 )
